Fix bug where yomichan wouldn't work on input box
Yomichan would only work on the input box if the auto-pasted text had
reached the text box. This was apparently happening because yomichan can
only process text that's within the <body>, and the body would only
extend to the text box when enough <p> elements containing auto-pasted
text had been inserted to overlap the body with the input box. Once this
occurred, yomichan would function as expected.