Bugfix: Adjust element layout
There was a bug where Yomichan would stop functioning once the page
scrolled, so I fixed this by having the <main> element be static with
its own scrollbar so that the page itself doesn't scroll. I couldn't
figure out any other way to fix this bug so I had to resort to this
which is kinda annoying and more of a workaround, but for a small
program like this, I think it's okay.

I also tweaked the style slightly in this commit.